requirements.txt
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.5
frozendict==2.4.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
lxml==5.1.0
MarkupSafe==2.1.3
multitasking==0.0.11
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pandas-ta==0.3.14b0
peewee==3.17.0
polygon-api-client==1.13.4
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
soupsieve==2.5
swagger-ui-bundle==0.0.9
tzdata==2023.4
urllib3==1.26.18
webencodings==0.5.1
websockets==12.0
Werkzeug==2.2.3
yfinance==0.2.35
zipp==3.17.0
